* Make WebGUI work with any kind of LDAP URI (ldap, ldaps and ldapi).
The fix consists on making every connection use the URI _without_ the base.
This requires rebuilding the URI parameters using scheme, host and port as
taken from the original URI stored in WebGUI's or User Preferences,
passing it as_string() to Net::LDAP->new(), and stripping any other
parameter to the constructor.

This way, ldap:// and ldapi:// work "out of the box". In order to make
ldaps:// work, the administrator must have a proper CA signing the
SSL certificate, and have it referenced in /etc/ldap/ldap.conf

In Auth/LDAP.pm:

  Since there are four connections, a private wrapper function was written.

In LDAPLink.pm

  Manual replacement of the single connection.

* Cosmetic. Remove 'use Net::LDAP' from SyncProfiles?ToLdap.pm since
  it is not being used at all.