Fixed WebGUI::Asset::Wobject::Search::view() so HTML highlighting not broken by search phrase with unmatched char pairs
Fixed WebGUI::Asset::Wobject::Search::view() so HTML highlighting will not
be broken by a search phrase with unmatched char pairs like '()', '{}',
and '[]' (the HTML highlighter uses a regex internally for matching words
in the search phrase). Also fixed so ':' in the search phrase does not
confuse the HTML highlighter.